Added an attribute for passing format specific custom extraction functions. The custom functions must be supplied like this:
{
	formatTag : [function1, function2, etc.],
	otherFormatTag : [function1, function2, etc.]
}

The two possible formatTag values are "OTF" and "WOFF".

The individual functions should accept two arguments representing the source font (A FontTools object)and the destination font (The font you passed as destinationFont). For example:

def myFunction(source, destination):
	blah

All of that said, this is untested, so please report any issues.